+++ src/parser/tests/parsing_failures.rs
@@ -205,30 +205,10 @@ x .foo
             use super::*;
 
             #[test]
-            fn block_without_indentation() {
-                let source = "
-foo: 42
-bar: 99
-";
-                check_parsing_fails(source);
-            }
-
-            #[test]
             fn block_starting_on_same_line_as_assignment() {
                 let source = "
 x = foo: 42
     bar: 99
-";
-                check_parsing_fails(source);
-            }
-
-            #[test]
-            fn block_without_indentation_in_function_after_first_line() {
-                let source = "
-f = ||
-  a = 1
-  foo: 42
-  bar: 99
 ";
                 check_parsing_fails(source);
             }
@@ -240,6 +220,17 @@ x =
   foo: 42
   bar
   baz: -1
+";
+                check_parsing_fails(source);
+            }
+
+            #[test]
+            fn block_value_with_multiple_expressions() {
+                let source = "
+x =
+  foo: 42
+  bar: 1, 2, 3
+  baz: 99
 ";
                 check_parsing_fails(source);
             }